requirements.txt
amqp==2.5.2
asgiref==3.2.7
backcall==0.1.0
billiard==3.6.3.0
celery==4.4.2
certifi==2020.4.5.1
chardet==3.0.4
decorator==4.4.2
dj-database-url==0.5.0
Django==3.0.5
django-debug-toolbar==2.2
django-extensions==2.2.9
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.0
ipdb==0.13.2
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.17.0
kombu==4.6.8
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
psycopg2==2.8.4
psycopg2-binary==2.8.5
ptyprocess==0.6.0
Pygments==2.6.1
pytz==2019.3
redis==3.4.1
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
traitlets==4.3.3
urllib3==1.25.9
virtualenv==16.7.9
vine==1.3.0
wcwidth==0.1.9
zipp==3.1.0
